+++ CHANGELOG.md
@@ -6,4 +6,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+
 - Initial release
+  - VS Code command to generate SSNs
+  - Highlight text on editor as an input
+    - Birthday
+    - Birthday + sex
+    - Age
+    - Age + sex